+++ elrond-wasm/src/storage/mod.rs
@@ -1,8 +1,6 @@
 pub mod storage_get;
 pub mod storage_set;
-pub mod mut_storage;
 pub mod protected_keys;
 
 pub use storage_get::*;
 pub use storage_set::*;
-pub use mut_storage::*;
