+++ elrond-wasm/src/types/mod.rs
@@ -1,5 +1,25 @@
-pub mod h256;
-pub mod queue;
+mod async_call_result;
+mod borrowed_mut_storage;
+mod h256;
+mod multi_args;
+mod multi_result;
+mod multi_result_vec;
+mod optional_arg;
+mod optional_result;
+mod queue;
+mod sc_error;
+mod sc_result;
+mod var_args;
 
-pub use h256::*;
-pub use queue::*;
+pub use async_call_result::{AsyncCallResult, AsyncCallError};
+pub use borrowed_mut_storage::BorrowedMutStorage;
+pub use h256::{H256, Address};
+pub use multi_args::*;
+pub use multi_result::*;
+pub use multi_result_vec::MultiResultVec;
+pub use optional_arg::OptionalArg;
+pub use optional_result::OptionalResult;
+pub use queue::Queue;
+pub use sc_error::SCError;
+pub use sc_result::SCResult;
+pub use var_args::VarArgs;
